ui/requirements.txt
streamlit==1.45.1
google-cloud-bigquery==3.31.0
db-dtypes==1.4.2
pandas==2.2.3
